MP3 encoder support 44K 22K 11K

+++ trunk/src/BleX264Encoder.hpp
@@ -45,7 +45,7 @@ public:
     int init();
     int encode(unsigned char *rgbframe, mint64 pts);
 
-    int getFrameDuration();
+    float getFrameDuration();
 
 private:
     void fini();
